+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/constants/FConstants.java
@@ -1,9 +1,7 @@
-package pedroPathing.constants;
+package org.firstinspires.ftc.teamcode.constants;
 
-import com.pedropathing.localization.Localizers;
 import com.pedropathing.follower.FollowerConstants;
-import com.pedropathing.util.CustomFilteredPIDFCoefficients;
-import com.pedropathing.util.CustomPIDFCoefficients;
+import com.pedropathing.localization.Localizers;
 import com.qualcomm.robotcore.hardware.DcMotorSimple;
 
 public class FConstants {
