+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/examples/ExampleFieldCentricTeleop.java
@@ -1,13 +1,13 @@
-package pedroPathing.examples;
+package org.firstinspires.ftc.teamcode.examples;
 
 import com.pedropathing.follower.Follower;
 import com.pedropathing.localization.Pose;
 import com.pedropathing.util.Constants;
-import  com.qualcomm.robotcore.eventloop.opmode.OpMode;
+import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 
-import pedroPathing.constants.FConstants;
-import pedroPathing.constants.LConstants;
+import org.firstinspires.ftc.teamcode.constants.FConstants;
+import org.firstinspires.ftc.teamcode.constants.LConstants;
 
 /**
  * This is an example teleop that showcases movement and field-centric driving.
