+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/tuners_tests/localization/SensorGoBildaPinpointExample.java
@@ -20,16 +20,15 @@
  *   SOFTWARE.
  */
 
-package pedroPathing.tuners_tests.localization;
+package org.firstinspires.ftc.teamcode.tuners_tests.localization;
 
-import com.qualcomm.robotcore.eventloop.opmode.Disabled;
+import com.pedropathing.localization.GoBildaPinpointDriver;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 
 import org.firstinspires.ftc.robotcore.external.navigation.AngleUnit;
 import org.firstinspires.ftc.robotcore.external.navigation.DistanceUnit;
 import org.firstinspires.ftc.robotcore.external.navigation.Pose2D;
-import com.pedropathing.localization.GoBildaPinpointDriver;
 
 import java.util.Locale;
